Move version range check to check-buildplan
