+++ src/it/check-buildplan-version-range-fail/invoker.properties
@@ -15,9 +15,5 @@
 # specific language governing permissions and limitations
 # under the License.
 
-# initial reference build: install
-invoker.goals.1=clean install
-# second build: verify (could be package, but not install to avoid overriding reference)
-invoker.goals.2=clean verify artifact:compare
-invoker.buildResult.2=failure
-
+invoker.goals=artifact:check-buildplan
+invoker.buildResult=failure
