+++ src/it/check-buildplan-version-range-fail/verify.groovy
@@ -20,6 +20,7 @@
 
 def buildLog = new File(basedir, 'build.log').text
 
-assert buildLog =~ /(?m)^\[WARNING] Dependency io\.cucumber:messages:jar:.+ \(compile\) via io\.cucumber:gherkin:jar:.+ is referenced with a range version \[32\.0\.0,33\.0\.0\)\s*$/
-assert buildLog =~ /(?m)^\[WARNING] Dependency commons-io:commons-io:jar:.+ \(compile\) is referenced with a range version \[2.20.0,\)\s*$/
-assert buildLog =~ /(?m)^\[WARNING] Dependency commons-collections:commons-collections:jar:LATEST \(compile\) is referenced with a range version LATEST\s*$/
+assert buildLog.contains('[ERROR] Version with range found in dependencies, project can be non-reproducible')
+assert buildLog.contains(' - Dependency io.cucumber:messages:jar:32.2.0 (compile) via io.cucumber:gherkin:jar:38.0.0 is referenced with a range version [32.0.0,33.0.0)')
+assert buildLog.contains(' - Dependency commons-io:commons-io:jar:2.21.0 (compile) is referenced with a range version [2.20.0,)')
+assert buildLog.contains(' - Dependency commons-collections:commons-collections:jar:LATEST (compile) is referenced with a range version LATEST')
