+++ src/it/check-buildplan-version-range-no-fail/invoker.properties
@@ -15,8 +15,4 @@
 # specific language governing permissions and limitations
 # under the License.
 
-# initial reference build: install
-invoker.goals.1=clean install
-# second build: verify (could be package, but not install to avoid overriding reference)
-invoker.goals.2=clean verify artifact:compare
-
+invoker.goals=artifact:check-buildplan
